+++ src/main/java/woowacourse/shoppingcart/dao/JdbcProductDao.java
@@ -1,0 +1,81 @@
+package woowacourse.shoppingcart.dao;
+
+import org.springframework.dao.EmptyResultDataAccessException;
+import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.jdbc.support.GeneratedKeyHolder;
+import org.springframework.security.core.parameters.P;
+import org.springframework.stereotype.Repository;
+import woowacourse.shoppingcart.domain.Product;
+import woowacourse.shoppingcart.exception.InvalidProductException;
+
+import java.sql.PreparedStatement;
+import java.util.List;
+import java.util.Objects;
+
+@Repository
+public class JdbcProductDao implements ProductDao {
+
+    private final JdbcTemplate jdbcTemplate;
+
+    public JdbcProductDao(final JdbcTemplate jdbcTemplate) {
+        this.jdbcTemplate = jdbcTemplate;
+    }
+
+    @Override
+    public Long save(final Product product) {
+        final String query = "INSERT INTO product (name, price, image_url, description, stock) VALUES (?, ?, ?, ?, ?)";
+        final GeneratedKeyHolder keyHolder = new GeneratedKeyHolder();
+        jdbcTemplate.update(connection -> {
+            final PreparedStatement preparedStatement =
+                    connection.prepareStatement(query, new String[]{"id"});
+            preparedStatement.setString(1, product.getName());
+            preparedStatement.setInt(2, product.getPrice());
+            preparedStatement.setString(3, product.getImageUrl());
+            preparedStatement.setString(4, product.getDescription());
+            preparedStatement.setInt(5, product.getStock());
+            return preparedStatement;
+        }, keyHolder);
+
+        return Objects.requireNonNull(keyHolder.getKey()).longValue();
+    }
+
+    @Override
+    public Product findProductById(final Long productId) {
+        try {
+            final String query = "SELECT name, price, image_url FROM product WHERE id = ?";
+            return jdbcTemplate.queryForObject(query, (resultSet, rowNumber) ->
+                    new Product(
+                            productId,
+                            resultSet.getString("name"),
+                            resultSet.getInt("price"),
+                            resultSet.getString("image_url"),
+                            resultSet.getString("description"),
+                            resultSet.getInt("image_url")
+                    ), productId
+            );
+        } catch (EmptyResultDataAccessException e) {
+            throw new InvalidProductException();
+        }
+    }
+
+    @Override
+    public List<Product> findProducts() {
+        final String query = "SELECT id, name, price, image_url FROM product";
+        return jdbcTemplate.query(query,
+                (resultSet, rowNumber) ->
+                        new Product(
+                                resultSet.getLong("id"),
+                                resultSet.getString("name"),
+                                resultSet.getInt("price"),
+                                resultSet.getString("image_url"),
+                                resultSet.getString("description"),
+                                resultSet.getInt("image_url")
+                        ));
+    }
+
+    @Override
+    public void delete(final Long productId) {
+        final String query = "DELETE FROM product WHERE id = ?";
+        jdbcTemplate.update(query, productId);
+    }
+}
