+++ src/test/java/woowacourse/shoppingcart/fixture/ProductFixtures.java
@@ -1,5 +1,7 @@
 package woowacourse.shoppingcart.fixture;
 
+import org.springframework.security.core.parameters.P;
+import woowacourse.shoppingcart.domain.Product;
 import woowacourse.shoppingcart.dto.ProductRequest;
 
 public class ProductFixtures {
@@ -12,5 +14,9 @@ public class ProductFixtures {
             "맛있는 라면", 1300,
             "https://user-images.githubusercontent.com/44823900/167772500-dff4dfb5-6ad2-48fe-937d-lamen.jpg",
             "개매움", 80);
+
+
+    public static final Product PRODUCT_1 = new Product(PRODUCT_REQUEST_1.getName(), PRODUCT_REQUEST_1.getPrice(),
+            PRODUCT_REQUEST_1.getImageUrl(), PRODUCT_REQUEST_1.getDescription(), PRODUCT_REQUEST_1.getStock());
 
 }
